Move reusable components to 'shared' directory.
