+++ src/components/shared/EuroCurrency.test.tsx
